requirements.txt
Flask
google-api-python-client
google-auth-httplib2
google-cloud-asset
google-cloud-logging